refactor: persona 模块按职责域垂直切分，新建基础设施
- 按 chat / life / gateway / tools / memory 重组目录结构
- 迁移 event_agent / scoring_agent / context_builder / observation_buffer /
  scheduler / delayed_task_queue / target_selector / models / utils /
  roll_adapter / privacy 到目标位置
- 旧位置放置 re-export shim (DeprecationWarning) 保持兼容
- 新建 ToolRegistry / MessagePort / MessagePipeline / ChatHooks / LifeHooks /
  HistoryManager 等基础设施骨架
- 同步更新 orchestrator.py / command.py / character_life.py 的 import 路径
- 合并 scheduler + delayed_task_queue 到 life/proactive.py
- 移除 observation_buffer 未使用的 max_records_per_group 参数
- 放宽 tool limit 类型检查兼容 LLM 返回的 float
