+++ src/plugins/DicePP/module/persona/gateway/pipeline.py
@@ -1,0 +1,52 @@
+"""消息处理管道 — 链式转换 SendAction 元数据，不做 I/O"""
+from typing import List
+from dataclasses import dataclass
+from abc import ABC, abstractmethod
+
+
+@dataclass
+class SendAction:
+    """发送动作元数据"""
+
+    user_id: str
+    group_id: str
+    content: str
+    delay_seconds: float = 0.0  # Pipeline 可设置, Port._execute_background 读取并执行
+
+
+class MessageStage(ABC):
+    """转换 SendAction 元数据, 不执行 I/O"""
+
+    @abstractmethod
+    async def process(self, actions: List[SendAction]) -> List[SendAction]:
+        ...
+
+
+class MessagePipeline:
+    """消息处理管道 — 链式转换 SendAction 元数据"""
+
+    def __init__(self):
+        self._stages: List[MessageStage] = []
+
+    def add(self, stage: MessageStage) -> "MessagePipeline":
+        self._stages.append(stage)
+        return self
+
+    async def process(self, actions: List[SendAction]) -> List[SendAction]:
+        result = actions
+        for stage in self._stages:
+            result = await stage.process(result)
+        return result
+
+
+class TruncateStage(MessageStage):
+    """超长消息截断"""
+
+    def __init__(self, max_chars: int = 2000):
+        self.max_chars = max_chars
+
+    async def process(self, actions: List[SendAction]) -> List[SendAction]:
+        for action in actions:
+            if len(action.content) > self.max_chars:
+                action.content = action.content[: self.max_chars - 3] + "..."
+        return actions
